# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Внутренний API дневника тренировок",
    "description": "API для доверенных сервисов. X-Api-Key проверяет вызывающий сервис; X-Telegram-User-Id выбирает пользователя. Не передавайте ключ клиентским приложениям: он позволяет действовать от имени любого пользователя. Вес в кг, расстояние в км, длительность в секундах.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "apiKey": [],
      "telegramIdentity": []
    }
  ],
  "paths": {
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "edit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/workouts": {
      "get": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/import": {
      "post": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "importCsv",
        "parameters": [
          {
            "name": "replaceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "allowDuplicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SetPlan": {
        "type": "object",
        "properties": {
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "warmup": {
            "type": "boolean"
          }
        }
      },
      "Target": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number"
          },
          "restSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "plan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetPlan"
            }
          }
        }
      },
      "TemplateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          }
        }
      },
      "TemplateItem": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "metricType": {
            "type": "string",
            "enum": [
              "STRENGTH",
              "BODYWEIGHT",
              "CARDIO",
              "TIMED"
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "targetSets": {
            "type": "integer",
            "format": "int32"
          },
          "targetReps": {
            "type": "integer",
            "format": "int32"
          },
          "targetWeight": {
            "type": "number"
          },
          "restSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "targetDuration": {
            "type": "integer",
            "format": "int32"
          },
          "targetDistance": {
            "type": "number"
          },
          "plan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetPlan"
            }
          }
        }
      },
      "TemplateView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateItem"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header"
      },
      "telegramIdentity": {
        "type": "apiKey",
        "name": "X-Telegram-User-Id",
        "in": "header"
      }
    }
  }
}